making the files more loosely coupled
