+++ api/internal/domain/Criterion.go
@@ -1,0 +1,7 @@
+package domain
+
+type Criterion struct {
+	ID          int    `json:"id"`
+	Name        string `json:"name"`
+	Explanation string `json:"explanation"`
+}
