+++ api/internal/domain/File.go
@@ -1,0 +1,6 @@
+package domain
+
+type File struct {
+	Path string `json:"path"`
+	// Add other fields as needed
+}
