+++ api/internal/domain/Item.go
@@ -1,0 +1,7 @@
+package domain
+
+type Item struct {
+	ID   string `json:"id"`
+	Name string `json:"name"`
+	// Add other fields as needed
+}
