+++ api/internal/domain/Submission.go
@@ -1,0 +1,5 @@
+package domain
+
+type Submission struct {
+	Criteria []Criterion `json:"criteria"`
+}
